+++ src/examples.js
@@ -16,7 +16,6 @@
 // components
 import * as AddImages from './components/AddImages/AddImages.example';
 import * as Avatar from './components/Avatar/Avatar.example';
-import * as BookingBreakdown from './components/BookingBreakdown/BookingBreakdown.example';
 import * as BookingDatesForm from './components/BookingPanel/BookingDatesForm/BookingDatesForm.example';
 import * as BookingPanel from './components/BookingPanel/BookingPanel.example';
 import * as BookingTimeInfo from './components/BookingTimeInfo/BookingTimeInfo.example';
@@ -63,6 +62,7 @@ import * as Menu from './components/Menu/Menu.example';
 import * as Modal from './components/Modal/Modal.example';
 import * as ModalInMobile from './components/ModalInMobile/ModalInMobile.example';
 import * as NamedLink from './components/NamedLink/NamedLink.example';
+import * as OrderBreakdown from './components/OrderBreakdown/OrderBreakdown.example';
 import * as OutsideClickHandler from './components/OutsideClickHandler/OutsideClickHandler.example';
 import * as PaginationLinks from './components/PaginationLinks/PaginationLinks.example';
 import * as PropertyGroup from './components/PropertyGroup/PropertyGroup.example';
@@ -118,7 +118,7 @@ export {
   ActivityFeed,
   AddImages,
   Avatar,
-  BookingBreakdown,
+  OrderBreakdown,
   BookingDateRangeFilter,
   BookingDatesForm,
   BookingTimeInfo,